Complete AC2-FIX-5 Claude compat internalization
