+++ crates/atm-core/src/boundary/store.rs
@@ -401,7 +401,11 @@ pub struct ConfigDoctorReport {
     pub findings: Vec<DoctorFinding>,
 }
 
-pub(crate) use atm_storage::compat::ProjectionAppendMode;
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "kebab-case")]
+pub enum ProjectionAppendMode {
+    RecoveredLogicalMessageSet,
+}
 
 /// Canonical non-Claude outbound request payload.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
